Add is_FindRecordShort method and update record usage
A new method `is_FindRecordShort` was added to `load_record.h` to check for the existence of records more efficiently. The `add_record_console.h` file was updated to use this new method instead of `is_FindRecord`. Additionally, a new user entry for "Doctor" Dan Health was added to `users.txt`.